+++ clang-tools-extra/test/clang-tidy/checkers/Inputs/Headers/unistd.h
@@ -1,4 +1,4 @@
-//===--- stdlib.h - Stub header for tests------ -----------------*- C++ -*-===//
+//===--- unistd.h - Stub header for tests------ -----------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,11 +6,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef _STDLIB_H_
-#define _STDLIB_H_
+#ifndef _UNISTD_H_
+#define _UNISTD_H_
 
-void abort(void);
-void _Exit(int);
-void quick_exit(int);
+void _exit(int status);
 
-#endif // _STDLIB_H_
+#endif // _UNISTD_H_